requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
coverage==7.6.4
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
libigl==2.5.1
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.2
psutil==6.1.0
pyhocon==0.3.61
pyparsing==3.2.0
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
sentry-sdk==2.17.0
setproctitle==1.3.3
setuptools==75.2.0
six==1.16.0
smmap==5.0.1
sympy==1.13.1
tifffile==2024.9.20
torch==2.5.0
torch-geometric==2.6.1
tqdm==4.66.5
trimesh==4.5.0
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
wandb==0.18.5
yarl==1.16.0
